backend/requirements.txt
Django==3.1.6
psycopg2==2.8.6
python-dotenv==0.15.0
djangorestframework==3.12.2
